refactor: simplify the input hander
Removed the worker thread and instead moved just the necessary functionality into `InputHandler` directly.
bench: 2577543
